Try to fix the simulator build warnings.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@12189 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/flac.c
@@ -450,8 +450,8 @@ enum codec_status codec_main(void)
     while (!*ci->taginfo_ready && !ci->stop_codec)
         ci->sleep(1);
     
-    ci->configure(DSP_SWITCH_FREQUENCY, (long *)(ci->id3->frequency));
-    ci->configure(DSP_SET_STEREO_MODE, (long *) (fc.channels == 1 
+    ci->configure(DSP_SWITCH_FREQUENCY, (int *)(ci->id3->frequency));
+    ci->configure(DSP_SET_STEREO_MODE, (int *) (fc.channels == 1 
         ? STEREO_MONO : STEREO_NONINTERLEAVED));
     codec_set_replaygain(ci->id3);
